- Updated Threshold type to float

+++ new-api/server/plugins/redis/models/agent.redis-model.js
@@ -28,7 +28,7 @@ const schema = {
         type: 'boolean'
     },
     domainClassifierThreshold: {
-        type: 'integer'
+        type: 'float'
     },
     fallbackResponses: {
         type: 'json'
